Automatic_water_meter_reading/requirements.txt
opencv-python
opencv-python-headless
flask
pillow
torch
torchvision
numpy
ultralytics